Move infer module to deploy folder to drop PyTorch dependency [for Docker mode]
